+++ js/index.js
@@ -9,7 +9,9 @@ import 'https://cdn.kernvalley.us/components/leaflet/map.js';
 import 'https://cdn.kernvalley.us/components/leaflet/marker.js';
 import 'https://cdn.kernvalley.us/components/pwa/install.js';
 import 'https://cdn.kernvalley.us/components/ad/block.js';
+import 'https://cdn.kernvalley.us/components/app/list-button.js';
 import { $, ready } from 'https://cdn.kernvalley.us/js/std-js/functions.js';
+import { init } from 'https://cdn.kernvalley.us/js/std-js/data-handlers.js';
 import { HTMLNotificationElement } from 'https://cdn.kernvalley.us/components/notification/html-notification.js';
 import { importGa, externalHandler, telHandler, mailtoHandler } from 'https://cdn.kernvalley.us/js/std-js/google-analytics.js';
 import { loadScript } from 'https://cdn.kernvalley.us/js/std-js/loader.js';
@@ -17,9 +19,39 @@ import { registerMapSearch } from './functions.js';
 import { hashChange, stateHandler } from './handlers.js';
 import { site, GA } from './consts.js';
 
-document.documentElement.classList.replace('no-js', 'js');
-document.documentElement.classList.toggle('no-dialog', document.createElement('dialog') instanceof HTMLUnknownElement);
-document.documentElement.classList.toggle('no-details', document.createElement('details') instanceof HTMLUnknownElement);
+$(document.documentElement).toggleClass({
+	'no-dialog': document.createElement('dialog') instanceof HTMLUnknownElement,
+	'no-details': document.createElement('details') instanceof HTMLUnknownElement,
+	'js': true,
+	'no-js': false,
+});
+
+cookieStore.get({ name: 'theme' }).then(async cookie => {
+	await $.ready;
+	const $ads = $('ad-block:not([theme]), ad-block[data-theme="auto"]');
+
+	const setTheme = async ({ name, value = 'auto' }) => {
+		if (name === 'theme') {
+			await Promise.all([
+				$(':root, [data-theme]').data({ theme: value }),
+				$('[theme]:not(ad-block)').attr({ theme: value }),
+				$ads.attr({ theme: value }),
+			]);
+		}
+	};
+
+	if (cookie && typeof cookie.value === 'string') {
+		setTheme(cookie);
+	}
+
+	cookieStore.addEventListener('change', ({ changed, deleted }) => {
+		const cookie = [...changed, ...deleted].find(({ name }) => name === 'theme');
+
+		if (cookie) {
+			setTheme(cookie);
+		}
+	});
+});
 
 if (typeof GA === 'string' && GA.length !== 0) {
 	requestIdleCallback(() => {
@@ -42,6 +74,8 @@ Promise.allSettled([
 	ready(),
 	loadScript('https://cdn.polyfill.io/v3/polyfill.min.js'),
 ]).then(async () => {
+	init().catch(console.error);
+
 	if (location.pathname === '/') {
 		requestIdleCallback(async () => {
 			registerMapSearch();
@@ -145,14 +179,6 @@ Promise.allSettled([
 		});
 	}
 
-	$('[data-scroll-to]').click(event => {
-		const target = document.querySelector(event.target.closest('[data-scroll-to]').dataset.scrollTo);
-		target.scrollIntoView({
-			bahavior: 'smooth',
-			block: 'start',
-		});
-	});
-
 	$('leaflet-marker').on('open', ({target}) => {
 		const url = new URL(location.pathname, location.origin);
 		url.hash = `#${target.id}`;
@@ -165,27 +191,6 @@ Promise.allSettled([
 				title: target.title,
 				uuid: target.id,
 			}, document.title, url.href);
-		}
-	});
-
-	$('[data-show]').click(event => {
-		const target = document.querySelector(event.target.closest('[data-show]').dataset.show);
-		if (target instanceof HTMLElement) {
-			target.show();
-		}
-	});
-
-	$('[data-show-modal]').click(event => {
-		const target = document.querySelector(event.target.closest('[data-show-modal]').dataset.showModal);
-		if (target instanceof HTMLElement) {
-			target.showModal();
-		}
-	});
-
-	$('[data-close]').click(event => {
-		const target = document.querySelector(event.target.closest('[data-close]').dataset.close);
-		if (target instanceof HTMLElement) {
-			target.tagName === 'DIALOG' ? target.close() : target.open = false;
 		}
 	});